+++ recipes/configure-engage-nginx-proxy.rb
@@ -8,10 +8,12 @@ include_recipe "oc-opsworks-recipes::install-nginx"
 shared_storage_root = get_shared_storage_root
 
 public_engage_hostname = get_public_engage_hostname
+public_engage_hostname_cors = get_public_engage_hostname_cors
+
 engage_whitelist = get_engage_admin_allowed_hosts
 # Use last 3 parts of Engage hostname domain for CORs domain check
 # Escape the 2 dots to use as literal dot chars in regex
-engage_domain_cors_regex = public_engage_hostname[/((?=.)(\w+))((?:.)(\w+)){2}\z/].gsub(/\./, "\\.")
+engage_domain_cors_regex = public_engage_hostname_cors[/((?=.)(\w+))((?:.)(\w+)){2}\z/].gsub(/\./, "\\.")
 
 ssl_info = node.fetch(:ssl, get_dummy_cert)
 if cert_defined(ssl_info)