+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## TO BE RELEASED
 
+* IC: OPC-851 `*.dcex.harvard.edu` cors support from Prod OC Engage
+
 ## v4.4.1
 
 * IC: `*.dcex.harvard.edu` cors support, requires nginx update on engage